+++ main/src/vocab/gpt2_vocab.rs
@@ -247,10 +247,10 @@ mod tests {
         let gpt2_vocab = Gpt2Vocab::from_file(path.to_path_buf().to_str().unwrap())?;
 
         //        When & Then
-        assert_eq!(gpt2_vocab.id_to_token(&(1 as i64)), "hello");
-        assert_eq!(gpt2_vocab.id_to_token(&(0 as i64)), "world");
-        assert_eq!(gpt2_vocab.id_to_token(&(3 as i64)), "!");
-        assert_eq!(gpt2_vocab.id_to_token(&(2 as i64)), "<|endoftext|>");
+        assert_eq!(gpt2_vocab.id_to_token(&(1_i64)), "hello");
+        assert_eq!(gpt2_vocab.id_to_token(&(0_i64)), "world");
+        assert_eq!(gpt2_vocab.id_to_token(&(3_i64)), "!");
+        assert_eq!(gpt2_vocab.id_to_token(&(2_i64)), "<|endoftext|>");
         drop(path);
         Ok(())
     }
